Soft Actor Critic and Applications (#1235)
* Soft Actor Critic and Applications

This commit adds SAC to garage. The version of
SAC added however is built for collecting batches
of transitions between optimization steps, and not
single transitions, as the original SAC does.

See #807

Furthermore small issues are addressed by this PR such
as #1229 and expanding the robustness for datatypes for
the TanhGaussianMLP Policy.

* Update torch/trpo_pendulum_ray_sampler.py ray init

- start the examples ray server in a lower resource
  mode.
- disable ray webui in CI

* Update Sac_half_cheetah_batch.py for more logging

- increase the number of logged epochs during training

+++ src/garage/torch/policies/tanh_gaussian_mlp_policy.py
@@ -109,7 +109,7 @@ class TanhGaussianMLPPolicy(Policy, GaussianMLPTwoHeadedModule):
         """
         with torch.no_grad():
             if not isinstance(observation, torch.Tensor):
-                observation = torch.from_numpy(observation).to(
+                observation = torch.from_numpy(observation).float().to(
                     tu.global_device())
             observation = observation.unsqueeze(0)
             dist = self.forward(observation)
@@ -137,7 +137,7 @@ class TanhGaussianMLPPolicy(Policy, GaussianMLPTwoHeadedModule):
         """
         with torch.no_grad():
             if not isinstance(observations, torch.Tensor):
-                observations = torch.from_numpy(observations).to(
+                observations = torch.from_numpy(observations).float().to(
                     tu.global_device())
             dist = self.forward(observations)
             ret_mean = dist.mean.cpu().numpy()
